% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg_exec.R
\name{ffmpeg_exec}
\alias{ffmpeg_exec}
\alias{have_ffmpeg_exec}
\title{Get Path to ffmpeg Executable}
\usage{
ffmpeg_exec()

have_ffmpeg_exec()
}
\value{
The path to the \code{ffmpeg} executable, or an error.
}
\description{
Get Path to ffmpeg Executable
}
\note{
This looks using \code{Sys.getenv("ffmpeg")} and \code{Sys.which("ffmpeg")}
to find \code{ffmpeg}.  If \code{ffmpeg} is not in your PATH, then please set the
path to \code{ffmpeg} using \code{Sys.setenv(ffmpeg = "/path/to/ffmpeg")}
}
\examples{
if (have_ffmpeg_exec()) { 
ffmpeg_exec()
}
}
