% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg_video_info.R
\name{ffmpeg_video_info}
\alias{ffmpeg_video_info}
\title{Video information from ffmpeg}
\usage{
ffmpeg_video_info(file)
}
\arguments{
\item{file}{Video file to get information from}
}
\value{
A character string
}
\description{
Video information from ffmpeg
}
\examples{
fname = system.file("extdata", "example.mp4", package = "ffmpeg")
if (have_ffmpeg_exec()) {
res = ffmpeg_video_info(fname)
}
}
